printf size_t

+++ import/decode_lzo.c
@@ -105,7 +105,7 @@ void decode_lzo(decode_t *decode)
 	if (verbose & TC_DEBUG) 
 	    fprintf (stderr, "got bytes (%ld)\n", bytes); 
 	if ( (ss=p_read (decode->fd_in, inbuf, bytes))!=bytes) {
-	    fprintf(stderr," (%s) failed to read frame: expected (%ld) got (%d)\n", __FILE__, bytes, ss);
+	    fprintf(stderr," (%s) failed to read frame: expected (%ld) got (%lu)\n", __FILE__, bytes, (unsigned long)ss);
 	    goto decoder_error;
 	}
 
@@ -124,7 +124,7 @@ void decode_lzo(decode_t *decode)
 	}
 
 	if ( (ss = p_write (decode->fd_out, out, out_len)) != out_len) {
-	    fprintf(stderr," (%s) failed to write frame: expected (%ld) wrote (%d)\n", __FILE__, bytes, ss);
+	    fprintf(stderr," (%s) failed to write frame: expected (%ld) wrote (%lu)\n", __FILE__, bytes, (unsigned long)ss);
 	    goto decoder_error;
 	}
     }
